refactor: move the component into the search folder
